UPBGE: Refactor texture renderers and fix planar and stereo.
One common issue of the texture renderer is the management of multiple
viewports (as for stereo). For renderer as cube map it doesn't matter
because none data from the viewport (view, projection) is used for the
cube faces, but planar does. Planar renderer use the camera of the
current viewport to figure out a mirror view matrix and use a projection
matrix near to the one of the viewport.

In case of stereo a planar renderer must be rendered twice for the both
stereo viewports. Previously it was achieved by separating the planar
in dependant renderer category and render this category in each viewport
render (RenderCamera function), but it implies to interrupt the render
of the viewport by unbinding its off screen.
Also in futur it would be interesting to create all the render data
and execute all the renders after without determining for who the
rasterizer is rendering.

At first point, the planars will all be rendered before the viewport
and use unique textures that will be selected in the viewport rendering.
Each of these textures are stored by layers (RAS_TextureRenderer::Layer)
which contain the render faces (Face class). Two situations are
possible:
- the renderer shares its layers along viewport, only one layer is
  available.
- the renderer use a unique layer per viewport.

The function RAS_TextureRenderer::EnsureLayers constructs the layers
if needed and return the usage of the layers : shared or unique.

When a layer is created, its constructor create a new GPUTexture
using the bind code from GPU_create_gl_tex and passing it to
GPU_texture_from_bindcode. Finally the texture is attached to the
frame buffers of the faces.

The renderer manager proceed the rendering per renderer, then per layer,
then per face. The render data changed to support this system.
First of all the camera data are now stored in the scene render data
and not per frame data. Also they are organized per eyes and frame
data use a list of eyes to specify which cameras could be rendered
for a frame. Each camera render data stores an unique index, this
index is used as renderer layer index.
When rendering, the camera data index is passed to RAS_BucketManager::Renderbuckets
which is used in fonction PrepareBuckets and KX_BlenderMaterial::Prepare
calling for each texture ApplyRenderer. This last function, in case
of renderer, get the bind code of the renderer depending on the
viewport index and copy this bind code for the next texture binding.

Finally the texture users are not needed anymore in RAS_TextureRenderer
as the renderer now create themselves the textures and changing the
mipmapping or the anisotropic filtering explicitly call the recreation
of the textures now.